+++ projects/showassignment/dense_heads/__init__.py
@@ -1,0 +1,3 @@
+from .yolov5_head_assigner import YOLOv5HeadAssigner
+
+__all__ = ['YOLOv5HeadAssigner']
